Transport for London Colours

A named vector of hexadecimal colour values from Transport for London's colour standard. Includes corporate, mode-specific and safety colours, as well as London Underground line colours.

oy_cols(...)

Arguments

...
Provide either the names or indices of colours in the vector (see details).

Value

A named vector object.

Details

An empty call returns the whole vector. The colour names are as follows:

corporate: corporate_blue, corporate_red, corporate_grey, corporate_dark_grey, corporate_yellow, corporate_green, corporate_black, corporate_white

mode-specific: transport_for_london, line_elizabeth, london_overground, emirates_air_line, santander_cycles, london_river_services, london_underground, london_buses, london_dial_a_ride, taxi_private_hire, visitor_centre, london_coaches, docklands_light_railway, tfl_rail

safety: safety_blue, corporate_red, corporate_yellow, corporate_green

London Underground lines: line_bakerloo, line_hammersmith_city, line_piccadilly, line_central, line_jubilee, line_victoria, line_circle, line_metropolitan, line_waterloo_city, line_district, line_northern

other logos and identities: oyster_blue, oyster_cyan, legible_london_blue, legible_london_yellow

Examples

oy_cols()  # prints the vector

#>          corporate_blue           corporate_red          corporate_grey 
#>               "#0019A8"               "#DC241F"               "#868F98" 
#>     corporate_dark_grey        corporate_yellow         corporate_green 
#>               "#414B56"               "#FFCE00"               "#007229" 
#>         corporate_black         corporate_white    transport_for_london 
#>               "#000000"               "#FFFFFF"               "#0019A8" 
#>          line_elizabeth       london_overground            london_trams 
#>               "#9364CC"               "#EF7B10"               "#00BD19" 
#>       emirates_air_line        santander_cycles   london_river_services 
#>               "#DC241F"               "#DC241F"               "#00A0E2" 
#>      london_underground            london_buses      london_dial_a_ride 
#>               "#0019A8"               "#DC241F"               "#B727BF" 
#>       taxi_private_hire          visitor_centre          london_coaches 
#>               "#8480D7"               "#DC006B"               "#F1AB00" 
#> docklands_light_railway                tfl_rail             safety_blue 
#>               "#00AFAD"               "#0019A8"               "#0060A8" 
#>           line_bakerloo   line_hammersmith_city         line_piccadilly 
#>               "#B26300"               "#F4A9BE"               "#0019A8" 
#>            line_central            line_jubilee           line_victoria 
#>               "#DC241F"               "#A1A5A7"               "#0098D8" 
#>             line_circle       line_metropolitan      line_waterloo_city 
#>               "#FFD329"               "#9B0058"               "#93CEBA" 
#>           line_district           line_northern             oyster_blue 
#>               "#007D32"               "#000000"               "#001df2" 
#>             oyster_cyan     legible_london_blue   legible_london_yellow 
#>               "#00f2f2"                "061a30"               "#f0bb00" 

oy_cols("oyster_blue", "oyster_cyan")  # prints

#> oyster_blue oyster_cyan 
#>   "#001df2"   "#00f2f2" 

oy_cols(1:5)  # provide indices

#>      corporate_blue       corporate_red      corporate_grey corporate_dark_grey 
#>           "#0019A8"           "#DC241F"           "#868F98"           "#414B56" 
#>    corporate_yellow 
#>           "#FFCE00"